style(components): move `connect-button` to `/connect`
