+++ src/components/connect/button.spec.tsx
@@ -1,7 +1,7 @@
 import { fireEvent, render, screen } from '@testing-library/react'
 import { describe, expect, test, vi } from 'vitest'
 
-import { ConnectButton } from './connect-button'
+import { ConnectButton } from './button'
 
 describe('ConnectButton', () => {
   test('should render the button', () => {
