+++ src/components/connect/button.tsx
@@ -1,12 +1,14 @@
 import { Button, ButtonProps } from 'primereact/button'
+import { classNames } from 'primereact/utils'
 
 import { environment } from '~/config'
 
-export function ConnectButton(props: ButtonProps) {
+export function ConnectButton({ className, ...props }: ButtonProps) {
   return (
     <Button
       {...props}
       severity="success"
+      className={classNames('justify-content-center flex', className)}
       onClick={() => window.open(`${environment.API_URL}/auth/login`, '_self')}
     >
       Connect
